implement json spell checker
