+++ packages/core/index.js
@@ -20,7 +20,6 @@ module.exports.security = coreAssertions.checkSecurity
 module.exports.checkUriSecurity = coreAssertions.checkUriSecurity
 module.exports.compress = compress
 module.exports.decompress = decompress
-module.exports.checkTypos = checkTypos
 module.exports.checkTmOptionalPointer = coreAssertions.checkTmOptionalPointer
 module.exports.checkLinkedAffordances = coreAssertions.checkLinkedAffordances
 module.exports.checkLinkedStructure = coreAssertions.checkLinkedStructure
@@ -1106,345 +1105,6 @@ function compress(data) {
  */
 function decompress(data) {
     return lzs.decompressFromEncodedURIComponent(data)
-}
-
-// --------------------------------------------------
-
-// -------------------------------------------------- checkTypos
-
-const REF = "$ref"
-const PROPERTIES = "properties"
-const ADDITONAL_PROPERTIES = "additional_properties"
-const DATA_SCHEMA = "dataSchema"
-const PATH = "#/"
-const TYPO_LOOKUP_TABLE = createSchemaLookupTable(tdSchema)
-
-/**
- * Checks possible typos in a TD
- * @param {object} td The TD to apply typo check on
- * @returns List of possible typos where the typo consists of string value of typo itself and the message,
- * another string value, to be prompted to the user for the fix
- */
-function checkTypos(td) {
-    const typos = []
-
-    const lookupTable = TYPO_LOOKUP_TABLE
-    const searchDepth = 1
-    const searchPath = PATH
-    let tdJson = {}
-
-    try {
-        tdJson = JSON.parse(td)
-    } catch (err) {
-        return typos
-    }
-
-    searchTypos(typos, tdJson, lookupTable, searchDepth, searchPath)
-
-    return typos
-}
-
-/**
- * Searching typos on a specific path and depth
- * @param {Array} typos The list that typo objects are stored
- * @param {object} tdJson JSON object of the TD
- * @param {Map} lookupTable The map that stores paths and their available word list according to their path depth
- * @param {integer} searchDepth The integer that decides the depth of the typo check search
- * @param {string} searchPath The string that decided the path of the typo check search
- */
-function searchTypos(typos, tdJson, lookupTable, searchDepth, searchPath) {
-    for (const key in tdJson) {
-        if (tdJson.hasOwnProperty(key)) {
-            const pathMap = lookupTable.get(searchDepth)
-            const wordSet = pathMap.get(searchPath)
-
-            if (wordSet.has(key)) {
-                continue
-            }
-
-            wordSet.forEach(word => {
-                if (doesTypoExist(key, word)) {
-                    typos.push({
-                        word: key,
-                        message: `Did you mean ${word}?`
-                    })
-
-                    return
-                }
-            })
-        }
-    }
-}
-
-/**
- * Creates a lookup table using JSON schema
- * @param {object} jsonSchema JSON Schema to create a lookup table from
- * @returns The map that constructs lookup table for typo check using TD Schema
- */
-function createSchemaLookupTable(jsonSchema) {
-    const lookupTable = new Map()
-    const filteredLookupTable = new Map()
-
-    findPathsInSchema(lookupTable, jsonSchema, PATH)
-
-    lookupTable.forEach((value, key) => {
-        if (value.size > 0) {
-            const pathDepth = (key.match(/\//ig) || []).length
-
-            let pathDepthMap = filteredLookupTable.get(pathDepth)
-
-            if (pathDepthMap) {
-                pathDepthMap.set(key.replace(/^r/g, ''), value)
-                filteredLookupTable.set(pathDepth, pathDepthMap)
-            } else {
-                pathDepthMap = new Map()
-                pathDepthMap.set(key.replace(/^r/g, ''), value)
-                filteredLookupTable.set(pathDepth, pathDepthMap)
-            }
-        }
-    })
-
-    return filteredLookupTable
-}
-
-/**
- * Finds the paths under a parent path by parsing schema and adds them to a lookup table
- * @param {Map} lookupTable The map that stores the paths in the schema
- * @param {object} schema The schema to find the paths from
- * @param {string} path The parent path that search is going under
- */
-function findPathsInSchema(lookupTable, schema, path) {
-    const keys = new Set()
-
-    if (schema[REF]) {
-        if (path[0] === 'r' && schema[REF].includes(DATA_SCHEMA)) {
-            return
-        }
-
-        if (schema[REF].includes(DATA_SCHEMA)) {
-            path = 'r' + path
-        }
-
-        findPathsInSchema(getRefObjectOfSchema(tdSchema, schema[REF]), path)
-        return
-    }
-
-    if (schema.type === 'object') {
-        const properties = schema[PROPERTIES]
-        for (const key in properties) {
-            if (properties.hasOwnProperty(key)) {
-                if (key === REF) {
-                    if (path[0] === 'r' && properties[key].includes(DATA_SCHEMA)) {
-                        continue
-                    }
-
-                    if (properties[key].includes(DATA_SCHEMA)) {
-                        path = 'r' + path
-                    }
-
-                    findPathsInSchema(getRefObjectOfSchema(tdSchema, properties[key]), path)
-                    return
-                } else {
-                    findPathsInSchema(properties[key], `${path}${key}/`)
-                    keys.add(key)
-                }
-            }
-        }
-
-        const additionalProperties = schema[ADDITONAL_PROPERTIES]
-        for (const key in additionalProperties) {
-            if (additionalProperties.hasOwnProperty(key)) {
-                if (key === REF) {
-                    if (path[0] === 'r' && additionalProperties[key].includes(DATA_SCHEMA)) {
-                        continue
-                    }
-
-                    if (additionalProperties[key].includes(DATA_SCHEMA)) {
-                        path = 'r' + path
-                    }
-
-                    findPathsInSchema(getRefObjectOfSchema(tdSchema, additionalProperties[key]), `${path}*/`)
-                    return
-                }
-            }
-        }
-
-        putKeysToPath(lookupTable, path, keys)
-    }
-
-    if (schema.type === 'array') {
-        const items = schema.items
-
-        for (const item in items) {
-            if (items.hasOwnProperty(item)) {
-                if (item === REF) {
-                    if (path[0] === 'r' && items[item].includes(DATA_SCHEMA)) {
-                        continue
-                    }
-
-                    if (items[item].includes(DATA_SCHEMA)) {
-                        path = 'r' + path
-                    }
-
-                    findPathsInSchema(getRefObjectOfSchema(tdSchema, items[item]), path)
-                    return
-                }
-            }
-        }
-
-        putKeysToPath(lookupTable, path, keys)
-    }
-
-    for (const key in schema) {
-        if (schema.hasOwnProperty(key)) {
-            if (['allOf', 'oneOf', 'anyOf'].includes(key)) {
-                if (Array.isArray(schema[key])) {
-                    schema[key].forEach(element => {
-                        findPathsInSchema(element, path)
-                    })
-                }
-            }
-        }
-    }
-}
-
-/**
- * Stores the keys under a specific path
- * @param {Map} lookupTable The map that stores the paths in the schema
- * @param {string} path The path that is owner of the current keys
- * @param {Set} keys The set of keys that is going to be put
- */
-function putKeysToPath(lookupTable, path, keys) {
-    pathKeys = lookupTable.get(path)
-
-    if (pathKeys) {
-        const union = new Set(pathKeys)
-        keys.forEach(k => {
-            union.add(k)
-        })
-
-        lookupTable.set(path, union)
-    } else {
-        lookupTable.set(path, keys)
-    }
-}
-
-/**
- * Gets the reference object in the schema
- * @param {object} schema The object that represent the schema
- * @param {string} ref The reference value in the schema
- * @returns The reference object the ref maps to
- */
-function getRefObjectOfSchema(schema, ref) {
-    const splitRef = ref.split('/')
-    if (splitRef[0] !== '#') {
-        console.log('Parsing not implemented for between files')
-        return
-    }
-
-    let result = schema
-
-    for (let i = 1; i < splitRef.length; i++) {
-        result = result[splitRef[i]]
-    }
-
-    return result
-}
-
-// Minimum similarity value to be able to say that two words are similar
-const SIMILARITY_THRESHOLD = 0.85
-
-// Maximum value of length difference between two words
-const MAX_LENGTH_DIFFERENCE = 2
-
-/**
- * Checks whether typo exists or not by comparing similarity of the two words
- * @param {string} actual The property name of the TD entered by user
- * @param {string} desired The desired property name that is retrieved from TD Schema
- * @returns Boolean value that tell whether typo exists or not
- */
-function doesTypoExist(actual, desired) {
-    if (Math.abs(actual.length - desired.length) > MAX_LENGTH_DIFFERENCE) {
-        return false
-    }
-
-    const similarity = calculateSimilarity(actual, desired)
-    return similarity > SIMILARITY_THRESHOLD && similarity !== 1.0
-}
-
-/**
- * Similarity of words calculated using Jaro-Winkler algorithm
- * @param {string} actual The property name of the TD entered by user
- * @param {string} desired The desired propert name that is retrieved from TD Schema
- * @returns Similarity of value the two inputs
- */
-function calculateSimilarity(actual, desired) {
-    let m = 0
-
-    if (actual.length === 0 || desired.length === 0) {
-        return 0
-    }
-
-    if (actual === desired) {
-        return 1
-    }
-
-    const range = Math.floor(Math.max(actual.length, desired.length) / 2) - 1
-    const actualMatches = new Array(actual.length)
-    const desiredMatches = new Array(desired.length)
-
-    // check lower and upper bounds to find the matches
-    for (let i = 0; i < actual.length; i++) {
-        const lowerBound = (i >= range) ? i - range : 0
-        const upperBound = (i + range <= desired.length) ? (i + range) : (desired.length - 1)
-
-        for (let j = lowerBound; j <= upperBound; j++) {
-            if (actualMatches[i] !== true && desiredMatches[j] !== true && actual[i] === desired[j]) {
-                m++
-                actualMatches[i] = desiredMatches[j] = true
-                break
-            }
-        }
-    }
-
-    if (m === 0) {
-        return 0
-    }
-
-    let k = 0
-    let transpositionCount = 0
-
-    // count transpositions
-    for (let i = 0; i < actual.length; i++) {
-        if (actualMatches[i] === true) {
-            let j = 0
-            for (j = k; j < desired.length; j++) {
-                if (desiredMatches[j] === true) {
-                    k = j + 1
-                    break
-                }
-            }
-
-            if (actual[i] !== desired[j]) {
-                transpositionCount++
-            }
-        }
-    }
-
-    let similarity = ((m / actual.length) + (m / desired.length) + ((m - (transpositionCount / 2)) / m)) / 3
-    let l = 0
-    const p = 0.1
-
-    // strengthen the similarity if the words start with same letters
-    if (similarity < 0.7) {
-        while (actual[l] === desired[l] && l < 4) {
-            l++
-        }
-
-        similarity = similarity + l * p * (1 - similarity)
-    }
-
-    return similarity
 }
 
 /**
