+++ packages/json-spell-checker/examples/open-api/open-api-example.json
@@ -1,0 +1,126 @@
+{
+    "swagger": "2.0",
+    "info": {
+        "versi2on": "1.0.0",
+        "titles": "Swagger Petstore",
+        "descriptiaon": "A sample API that uses a petstore as an example to demonstrate features in the swagger-2.0 specification",
+        "termsOfService": "http://swagger.io/terms/",
+        "contact": {
+        "name": "Swagger API Team",
+        "email": "apiteam@swagger.io",
+        "url": "http://swagger.io"
+        },
+        "licensse": {
+        "name": "Apache 2.0",
+        "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
+        }
+    },
+    "host": "petstore.swagger.io",
+    "basePath": "/api",
+    "schemes2": [
+        "http"
+    ],
+    "consumes": [
+        "application/json"
+    ],
+    "prodsuces": [
+        "application/json"
+    ],
+    "paths": {
+        "/pets": {
+            "get": {
+                "descriptison": "Returns all pets from the system that the user has access to\nNam sed condimentum est. Maecenas tempor sagittis sapien, nec rhoncus sem sagittis sit amet. Aenean at gravida augue, ac iaculis sem. Curabitur odio lorem, ornare eget elementum nec, cursus id lectus. Duis mi turpis, pulvinar ac eros ac, tincidunt varius justo. In hac habitasse platea dictumst. Integer at adipiscing ante, a sagittis ligula. Aenean pharetra tempor ante molestie imperdiet. Vivamus id aliquam diam. Cras quis velit non tortor eleifend sagittis. Praesent at enim pharetra urna volutpat venenatis eget eget mauris. In eleifend fermentum facilisis. Praesent enim enim, gravida ac sodales sed, placerat id erat. Suspendisse lacus dolor, consectetur non augue vel, vehicula interdum libero. Morbi euismod sagittis libero sed lacinia.\n\nSed tempus felis lobortis leo pulvinar rutrum. Nam mattis velit nisl, eu condimentum ligula luctus nec. Phasellus semper velit eget aliquet faucibus. In a mattis elit. Phasellus vel urna viverra, condimentum lorem id, rhoncus nibh. Ut pellentesque posuere elementum. Sed a varius odio. Morbi rhoncus ligula libero, vel eleifend nunc tristique vitae. Fusce et sem dui. Aenean nec scelerisque tortor. Fusce malesuada accumsan magna vel tempus. Quisque mollis felis eu dolor tristique, sit amet auctor felis gravida. Sed libero lorem, molestie sed nisl in, accumsan tempor nisi. Fusce sollicitudin massa ut lacinia mattis. Sed vel eleifend lorem. Pellentesque vitae felis pretium, pulvinar elit eu, euismod sapien.\n",
+                "operationId": "findPets",
+                "parametsers": [
+                {
+                    "name": "tags",
+                    "in": "query",
+                    "description": "tags to filter by",
+                    "required": false,
+                    "type": "array",
+                    "collectionFormat": "csv",
+                    "items": {
+                    "type": "string"
+                    }
+                },
+                {
+                    "name": "limit",
+                    "in": "query",
+                    "description": "maximum number of results to return",
+                    "required": false,
+                    "type": "integer",
+                    "format": "int32"
+                }
+                ],
+                "responses": {
+                "200": {
+                    "description": "pet response",
+                    "schema": {
+                    "type": "array",
+                    "items": {
+                        "$ref": "#/definitions/Pet"
+                    }
+                    }
+                },
+                "default": {
+                    "description": "unexpected error",
+                    "schema": {
+                    "$ref": "#/definitions/Error"
+                    }
+                }
+                }
+            },
+            "posst": {
+                "description": "Creates a new pet in the store.  Duplicates are allowed",
+                "operationId": "addPet",
+                "parameters": [
+                {
+                    "name": "pet",
+                    "in": "body",
+                    "description": "Pet to add to the store",
+                    "required": true,
+                    "schema": {
+                    "$ref": "#/definitions/NewPet"
+                    }
+                }
+                ],
+                "responses": {
+                "200": {
+                    "description": "pet response",
+                    "schema": {
+                    "$ref": "#/definitions/Pet"
+                    }
+                },
+                "default": {
+                    "description": "unexpected error",
+                    "schema": {
+                    "$ref": "#/definitions/Error"
+                    }
+                }
+                }
+            }
+        }
+    },
+    "definitions": {
+        "Pet": {
+            "type": "object",
+            "allOf": [
+                {
+                    "$ref": "#/definitions/NewPet"
+                },
+                {
+                    "required": [
+                        "id"
+                    ],
+                    "properties": {
+                        "id": {
+                        "type": "integer",
+                        "format": "int64"
+                        }
+                    }
+                }
+            ]
+        }  
+    },
+    "typoCount": 7
+}
