+++ packages/json-spell-checker/examples/td/td-example.json
@@ -1,0 +1,97 @@
+{
+    "id2": "urn:formOpArrayWithDefaults",
+	"@co3ntext": "https://www.w3.org/2022/wot/td/v1.1",
+	"titlve": "MyLampThing",
+	"descbription": "Valid TD demonstrating how to use multiple forms with different op",
+	"securnityDefinitions": {
+		"basic_sc": {
+			"scheme": "basic",
+			"in": "header"
+		}
+	},
+	"securimty": [
+		"basic_sc"
+	],
+	"properties": {
+		"brightness": {
+			"ddescription": "The current brightness setting",
+			"tyfpe": "integer",
+			"mingimum": -64,
+			"maximum": 64,
+			"observable": true,
+			"readOnly": false,
+			"writeOnly": false,
+			"forms": [
+				{
+					"hresf": "http://example.org:9191/api/brightness",
+					"op": [
+						"readproperty"
+					],
+					"contentType": "application/json"
+				},
+				{
+					"href": "http://example.org:9191/api/brightness",
+					"op": [
+						"writeproperty"
+					],
+					"htv:methodName": "POST",
+					"contentType": "application/json"
+				},
+				{
+					"href": "http://example.org:9191/api/brightness/observe",
+					"op": [
+						"observeproperty"
+					],
+					"htv:methodName": "GET",
+					"subprotocol": "websub",
+					"contentType": "application/json"
+				},
+				{
+					"href": "http://example.org:9191/api/brightness/unobserve",
+					"op": [
+						"unobserveproperty"
+					],
+					"htv:methodName": "GET",
+					"subprotocol": "websub",
+					"contentType": "application/json"
+				}
+			]
+		}
+	},
+	"actions": {
+		"toggle": {
+			"safe": false,
+			"idempotent": false,
+			"forms": [
+				{
+					"href": "https://mylamp.example.com/toggle",
+					"op": [
+						"invokeaction"
+					],
+					"contentType": "application/json"
+				}
+			]
+		}
+	},
+	"events": {
+		"overheating": {
+			"data": {
+				"type": "string",
+				"readOnsly": true,
+				"writeOnly": false,
+				"observable": false
+			},
+			"forms": [
+				{
+					"href": "https://mylamp.example.com/oh",
+					"subprotocol": "longpoll",
+					"op": [
+						"subscribeevent"
+					],
+					"contentType": "application/json"
+				}
+			]
+		}
+	},
+	"typoCount": 15
+}
