+++ packages/web/index.html
@@ -355,6 +355,7 @@
   <script src="./node_modules/@thing-description-playground/td_to_openapi/dist/web-bundle.min.js"></script>
   <script src="./node_modules/@thing-description-playground/td_to_asyncapi/dist/web-bundle.min.js"></script>
   <script src="./node_modules/@thing-description-playground/defaults/dist/web-bundle.min.js"></script>
+  <script src="./node_modules/@thing-description-playground/spell_checker/dist/web-bundle.min.js"></script>
   <!-- Ensure that d3 and d3-tip are added before loader.js,
     otherwise throws an error, see https://github.com/microsoft/monaco-editor/issues/2283 -->
   <!-- Also order matters for svg-crowbar, vega and vega-embed (relative to loader.js) -->
